Various improvements

+++ review/static_src/src/types.ts
@@ -9,16 +9,49 @@ export class CommentReply {
     }
 }
 
-class BaseComment {
-    id: number;
+type CommentState = 'default' | 'creating' | 'editing' | 'saving' | 'deleting' | 'deleted' | 'save_error' | 'delete_error';
+
+export class Comment {
+    localId: number;
     annotation: Annotation;
+    remoteId: number|null;
+    state: CommentState;
+    author: string;
+    text: string;
+    replies: CommentReply[];
+    newReply: string;
     _element: HTMLElement|null;
 
-    constructor(id: number, annotation: Annotation) {
-        this.id = id;
+    constructor(localId: number, annotation: Annotation, {remoteId=null, state=<CommentState>'default', author='', text='', replies=[], newReply=''}) {
+        this.localId = localId;
         this.annotation = annotation;
+        this.remoteId = remoteId;
+        this.state = state;
+        this.author = author;
+        this.text = text;
+        this.replies = replies;
+        this.newReply = newReply;
 
         this._element = null;
+    }
+
+    static makeNew(localId: number, annotation: Annotation): Comment {
+        return new Comment(localId, annotation, {state: 'creating'});
+    }
+
+    static fromApi(localId: number, annotation: Annotation, data: any): Comment {
+        return new Comment(localId, annotation, {remoteId: data.id, author: data.author, text: data.text});
+    }
+
+    clone() {
+        return new Comment(this.localId, this.annotation, {
+            remoteId: this.remoteId,
+            state: this.state,
+            author: this.author,
+            text: this.text,
+            replies: this.replies,
+            newReply: this.newReply,
+        });
     }
 
     getDesiredPosition(): number {
@@ -29,64 +62,36 @@ class BaseComment {
 
     getElement(containerElement: HTMLElement): HTMLElement {
         if (!this._element) {
-            this._element = document.querySelector(`[data-comment-id="${this.id}"]`);
+            this._element = document.querySelector(`[data-comment-id="${this.localId}"]`);
         }
 
         return this._element;
     }
-}
-
-export class Comment extends BaseComment {
-    remoteId: number;
-    author: string;
-    text: string;
-    isDeleting: boolean;
-    replies: CommentReply[];
-    replyInProgress: string;
-
-    constructor(id: number, data: any, annotation: Annotation) {
-        super(id, annotation);
-
-        this.remoteId = data.id;
-        this.author = "AUTHOR";
-        this.text = data.text;
-        this.isDeleting = false;
-        this.replies = [];
-        this.replyInProgress = '';
-    }
 
     getDateDisplay() {
-        return "DATE";
+        return "FOO";
     }
 
     async delete(api: APIClient) {
-        this.isDeleting = true;
+        this.state = 'deleting';
 
-        await fetch(`${api.baseUrl}/comments/${this.remoteId}/`, {
-            method: 'DELETE',
-            headers: {
-                'X-Review-Token': api.reviewToken,
-            },
-        });
+        if (this.remoteId) {
+            await fetch(`${api.baseUrl}/comments/${this.remoteId}/`, {
+                method: 'DELETE',
+                headers: {
+                    'X-Review-Token': api.reviewToken,
+                },
+            });
+        }
 
         this.annotation.onDelete();
-    }
-}
-
-export class NewComment extends BaseComment {
-    text: string;
-    isNew: boolean;
-    isSubmitting: boolean;
-
-    constructor(id: number, annotation: Annotation) {
-        super(id, annotation);
-        this.text = "";
-        this.isNew = true;
-        this.isSubmitting = false;
+        this.state = 'deleted';
     }
 
-    async submit(api: APIClient) {
-        this.isSubmitting = true;
+    async save(api: APIClient) {
+        this.state = 'saving';
+
+        // TODO: PUT if has remote id
 
         let response = await fetch(`${api.baseUrl}/comments/`, {
             method: 'POST',
@@ -106,10 +111,11 @@ export class NewComment extends BaseComment {
             }),
         });
 
-        return new Comment(this.id, await response.json(), this.annotation);
-    }
+        let commentData = await response.json();
 
-    cancel() {
-        this.annotation.onDelete();
+        this.state = 'default';
+        this.remoteId = commentData.id;
+
+        // TODO: Set other attributes from response
     }
 }
